move pdg to shower package
