+++ grand/simu/acqui_rf.py
