+++ grand/simu/shower/__init__.py
@@ -1,5 +1,6 @@
 from .coreas import CoreasShower
 from .generic import ShowerEvent
 from .zhaires import ZhairesShower
+from .pdg import ParticleCode
 
-__all__ = ["CoreasShower", "ShowerEvent", "ZhairesShower"]
+__all__ = ["CoreasShower", "ShowerEvent", "ZhairesShower","ParticleCode"]
