+++ grand/simu/shower/coreas.py
@@ -10,7 +10,7 @@ import numpy
 from grand import CartesianRepresentation, SphericalRepresentation
 from grand.simu.shower.generic import CollectionEntry, FieldsCollection, ShowerEvent
 from grand.simu.antenna import ElectricField
-from grand.simu.pdg import ParticleCode
+from grand.simu.shower.pdg import ParticleCode
 
 __all__ = ["CoreasShower"]
